refactor(Fixture): 도메인별로 fixture 나누도록 함
