+++ src/test/java/kitchenpos/acceptance/MenuAcceptanceTest.java
@@ -2,10 +2,6 @@ package kitchenpos.acceptance;
 
 import static kitchenpos.acceptance.MenuGroupAcceptanceTest.메뉴그룹_등록_요청;
 import static kitchenpos.acceptance.ProductAcceptanceTest.상품_등록_요청;
-import static kitchenpos.fixture.DomainFactory.createMenu;
-import static kitchenpos.fixture.DomainFactory.createMenuGroup;
-import static kitchenpos.fixture.DomainFactory.createMenuProduct;
-import static kitchenpos.fixture.DomainFactory.createProduct;
 import static kitchenpos.fixture.MenuFactory.createMenuRequest;
 import static kitchenpos.fixture.MenuProductFactory.createMenuProductRequest;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -18,16 +14,10 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Price;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductName;
 import kitchenpos.dto.MenuGroupResponse;
 import kitchenpos.dto.MenuProductRequest;
 import kitchenpos.dto.MenuRequest;
 import kitchenpos.dto.MenuResponse;
-import kitchenpos.dto.ProductRequest;
 import kitchenpos.dto.ProductResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
