+++ src/test/java/kitchenpos/acceptance/OrderAcceptanceTest.java
@@ -4,13 +4,6 @@ import static kitchenpos.acceptance.MenuAcceptanceTest.메뉴_등록_요청;
 import static kitchenpos.acceptance.MenuGroupAcceptanceTest.메뉴그룹_등록_요청;
 import static kitchenpos.acceptance.ProductAcceptanceTest.상품_등록_요청;
 import static kitchenpos.acceptance.TableAcceptanceTest.주문테이블_등록_요청;
-import static kitchenpos.fixture.DomainFactory.createMenu;
-import static kitchenpos.fixture.DomainFactory.createMenuGroup;
-import static kitchenpos.fixture.DomainFactory.createMenuProduct;
-import static kitchenpos.fixture.DomainFactory.createOrder;
-import static kitchenpos.fixture.DomainFactory.createOrderLineItem;
-import static kitchenpos.fixture.DomainFactory.createOrderTable;
-import static kitchenpos.fixture.DomainFactory.createProduct;
 import static kitchenpos.fixture.MenuProductFactory.createMenuProductRequest;
 import static kitchenpos.fixture.OrderFactory.createOrderLineItemRequest;
 import static kitchenpos.fixture.OrderFactory.createOrderRequest;
@@ -24,12 +17,7 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.Product;
 import kitchenpos.dto.MenuGroupResponse;
 import kitchenpos.dto.MenuResponse;
 import kitchenpos.dto.OrderLineItemRequest;
