+++ src/test/java/kitchenpos/acceptance/ProductAcceptanceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.acceptance;
 
-import static kitchenpos.fixture.DomainFactory.createProduct;
+import static kitchenpos.fixture.ProductFactory.createProduct;
 import static kitchenpos.fixture.ProductFactory.createProductRequest;
 import static org.assertj.core.api.Assertions.assertThat;
