+++ src/test/java/kitchenpos/acceptance/TableAcceptanceTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.acceptance;
 
-import static kitchenpos.fixture.DomainFactory.createOrderTable;
 import static kitchenpos.fixture.OrderTableFactory.createOrderTableRequest;
 import static org.assertj.core.api.Assertions.assertThat;
