+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import static kitchenpos.fixture.DomainFactory.createOrderTable;
+import static kitchenpos.fixture.OrderTableFactory.createOrderTable;
 import static kitchenpos.fixture.OrderTableFactory.createOrderTableRequest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
